schedular/requirements.txt
celery>=5.3.0
redis>=4.5.0
pymongo>=4.14.0
pytz>=2023.3
requests>=2.31.0
python-dotenv>=1.1.0
celery-redbeat